+++ src/App.jsx
@@ -29,7 +29,7 @@ import MainPage from "./pages/MainPage";
 import SearchPage from "./pages/SearchPage";
 import DeleteAccountPage from "./pages/DeleteAccountPage";
 import ReviewWriter from "./components/review-page/ReviewWriter";
-import { saveData } from "./constants/save";
+import { saveData } from "./mock/saveData";
 import { CampingDaysProvider } from "./context/campingDaysContext";
 import { AuthProvider } from "./context/authContext";
