+++ src/constants/cash.jsx
@@ -1,24 +1,4 @@
 /*****************************
- * 	캐시 충전/사용 데이터
- */
-export const cashData = [
-  {
-    cashStatus: "충전",
-    createdAt: "24-09-15",
-    transactionLog: "캐시 충전",
-    cash: 100000,
-    totalCash: 120000,
-  },
-  {
-    cashStatus: "사용",
-    createdAt: "24-08-09",
-    transactionLog: "파이브이모션",
-    cash: 80000,
-    totalCash: 20000,
-  },
-];
-
-/*****************************
  * 	캐시 충전 모달
  */
 export const chargePrice = [
