+++ src/constants/mypage.jsx
@@ -1,12 +1,4 @@
 /*****************************
- * 	유저 정보 상수데이터
- */
-export const myPageData = {
-  loginId: "camper123",
-  userImg: "",
-};
-
-/*****************************
  * 	버튼 데이터
  */
 export const mypageBtns = [
