+++ src/mock/mypageData.jsx
@@ -1,0 +1,7 @@
+/*****************************
+ * 	유저 정보 상수데이터
+ */
+export const myPageData = {
+  loginId: "camper123",
+  userImg: "",
+};
