+++ src/pages/AdminBookListPage.jsx
@@ -4,7 +4,7 @@ import { ReactSVG } from "react-svg";
 import { campBookData } from "../mock/campBookData"; // 목업 데이터 가져오기
 import { filterData } from "../utils/filterData";
 import { useLoadMore } from "../hooks/useLoadMore";
-import { noticeData } from "../constants/notice";
+import { noticeData } from "../mock/noticeData";
 import SaveMoreBtn from "../components/save-page/SaveMoreBtn";
 import "../components/admin-book-page/AdminBookListPage.css";
